move class on execution to core module
