+++ libs/idp-server-authentication-interactors/src/main/java/org/idp/server/authentication/interactors/email/executor/EmailChallengeAuthenticationExecutor.java
@@ -17,9 +17,6 @@
 package org.idp.server.authentication.interactors.email.executor;
 
 import java.util.Map;
-import org.idp.server.authentication.interactors.AuthenticationExecutionRequest;
-import org.idp.server.authentication.interactors.AuthenticationExecutionResult;
-import org.idp.server.authentication.interactors.AuthenticationExecutor;
 import org.idp.server.authentication.interactors.email.EmailAuthenticationConfiguration;
 import org.idp.server.authentication.interactors.email.EmailVerificationTemplate;
 import org.idp.server.authentication.interactors.email.OneTimePassword;
@@ -27,6 +24,9 @@ import org.idp.server.authentication.interactors.email.OneTimePasswordGenerator;
 import org.idp.server.authentication.interactors.sms.executor.SmsVerificationChallenge;
 import org.idp.server.core.openid.authentication.AuthenticationTransactionIdentifier;
 import org.idp.server.core.openid.authentication.config.AuthenticationExecutionConfig;
+import org.idp.server.core.openid.authentication.interaction.execution.AuthenticationExecutionRequest;
+import org.idp.server.core.openid.authentication.interaction.execution.AuthenticationExecutionResult;
+import org.idp.server.core.openid.authentication.interaction.execution.AuthenticationExecutor;
 import org.idp.server.core.openid.authentication.repository.AuthenticationInteractionCommandRepository;
 import org.idp.server.core.openid.authentication.repository.AuthenticationInteractionQueryRepository;
 import org.idp.server.platform.json.JsonConverter;
