+++ libs/idp-server-authentication-interactors/src/main/java/org/idp/server/authentication/interactors/fidouaf/FidoUafRegistrationInteractor.java
@@ -19,15 +19,15 @@ package org.idp.server.authentication.interactors.fidouaf;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import org.idp.server.authentication.interactors.AuthenticationExecutionRequest;
-import org.idp.server.authentication.interactors.AuthenticationExecutionResult;
-import org.idp.server.authentication.interactors.AuthenticationExecutor;
-import org.idp.server.authentication.interactors.AuthenticationExecutors;
 import org.idp.server.authentication.interactors.fidouaf.plugin.FidoUafAdditionalRequestResolvers;
 import org.idp.server.core.openid.authentication.*;
 import org.idp.server.core.openid.authentication.config.AuthenticationConfiguration;
 import org.idp.server.core.openid.authentication.config.AuthenticationExecutionConfig;
 import org.idp.server.core.openid.authentication.config.AuthenticationInteractionConfig;
+import org.idp.server.core.openid.authentication.interaction.execution.AuthenticationExecutionRequest;
+import org.idp.server.core.openid.authentication.interaction.execution.AuthenticationExecutionResult;
+import org.idp.server.core.openid.authentication.interaction.execution.AuthenticationExecutor;
+import org.idp.server.core.openid.authentication.interaction.execution.AuthenticationExecutors;
 import org.idp.server.core.openid.authentication.policy.AuthenticationPolicy;
 import org.idp.server.core.openid.authentication.repository.AuthenticationConfigurationQueryRepository;
 import org.idp.server.core.openid.authentication.repository.AuthenticationInteractionQueryRepository;
