+++ libs/idp-server-authentication-interactors/src/main/java/org/idp/server/authentication/interactors/webauthn/WebAuthnRegistrationIChallengeInteractorFactory.java
@@ -16,8 +16,8 @@
 
 package org.idp.server.authentication.interactors.webauthn;
 
-import org.idp.server.authentication.interactors.AuthenticationExecutors;
 import org.idp.server.core.openid.authentication.AuthenticationInteractor;
+import org.idp.server.core.openid.authentication.interaction.execution.AuthenticationExecutors;
 import org.idp.server.core.openid.authentication.plugin.AuthenticationDependencyContainer;
 import org.idp.server.core.openid.authentication.plugin.AuthenticationInteractorFactory;
 import org.idp.server.core.openid.authentication.repository.AuthenticationConfigurationQueryRepository;
