+++ libs/idp-server-authentication-interactors/src/main/resources/META-INF/services/org.idp.server.core.openid.authentication.interaction.execution.AuthenticationExecutorFactory
@@ -1,5 +1,5 @@
-org.idp.server.authentication.interactors.HttpRequestAuthenticationExecutorFactory
-org.idp.server.authentication.interactors.HttpRequestsAuthenticationExecutorFactory
+org.idp.server.core.openid.authentication.interaction.execution.HttpRequestAuthenticationExecutorFactory
+org.idp.server.core.openid.authentication.interaction.execution.HttpRequestsAuthenticationExecutorFactory
 org.idp.server.authentication.interactors.sms.executor.SmsChallengeAuthenticationExecutorFactory
 org.idp.server.authentication.interactors.sms.executor.SmsAuthenticationExecutorFactory
 org.idp.server.authentication.interactors.email.executor.EmailChallengeAuthenticationExecutorFactory
