+++ libs/idp-server-core/src/main/java/org/idp/server/core/openid/authentication/interaction/execution/AuthenticationExecutionRequest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.idp.server.authentication.interactors;
+package org.idp.server.core.openid.authentication.interaction.execution;
 
 import java.util.HashMap;
 import java.util.Map;
