+++ libs/idp-server-core/src/main/java/org/idp/server/core/openid/authentication/interaction/execution/AuthenticationExecutorFactory.java
@@ -14,32 +14,11 @@
  * limitations under the License.
  */
 
-package org.idp.server.authentication.interactors;
+package org.idp.server.core.openid.authentication.interaction.execution;
 
-public enum AuthenticationExecutionStatus {
-  OK(200),
-  CLIENT_ERROR(400),
-  SERVER_ERROR(500);
+import org.idp.server.core.openid.authentication.plugin.AuthenticationDependencyContainer;
 
-  int statusCode;
+public interface AuthenticationExecutorFactory {
 
-  AuthenticationExecutionStatus(int statusCode) {
-    this.statusCode = statusCode;
-  }
-
-  public boolean isOk() {
-    return this == OK;
-  }
-
-  public boolean isClientError() {
-    return this == CLIENT_ERROR;
-  }
-
-  public boolean isServerError() {
-    return this == SERVER_ERROR;
-  }
-
-  public int code() {
-    return statusCode;
-  }
+  AuthenticationExecutor create(AuthenticationDependencyContainer container);
 }
