+++ libs/idp-server-springboot-adapter/src/main/java/org/idp/server/adapters/springboot/application/restapi/metadata/FidoUafDiscoveryV1Api.java
@@ -19,8 +19,8 @@ package org.idp.server.adapters.springboot.application.restapi.metadata;
 import jakarta.servlet.http.HttpServletRequest;
 import org.idp.server.IdpServerApplication;
 import org.idp.server.adapters.springboot.application.restapi.ParameterTransformable;
-import org.idp.server.authentication.interactors.AuthenticationMetadataResponse;
 import org.idp.server.authentication.interactors.fidouaf.AuthenticationMetaDataApi;
+import org.idp.server.core.openid.authentication.interaction.execution.AuthenticationMetadataResponse;
 import org.idp.server.platform.multi_tenancy.tenant.TenantIdentifier;
 import org.idp.server.platform.type.RequestAttributes;
 import org.springframework.http.HttpHeaders;
