+++ libs/idp-server-use-cases/src/main/java/org/idp/server/usecases/application/enduser/AuthenticationMetaDataEntryService.java
@@ -17,13 +17,13 @@
 package org.idp.server.usecases.application.enduser;
 
 import java.util.Map;
-import org.idp.server.authentication.interactors.*;
 import org.idp.server.authentication.interactors.fidouaf.*;
 import org.idp.server.core.openid.authentication.AuthenticationTransactionIdentifier;
 import org.idp.server.core.openid.authentication.config.AuthenticationConfiguration;
 import org.idp.server.core.openid.authentication.config.AuthenticationExecutionConfig;
 import org.idp.server.core.openid.authentication.config.AuthenticationInteractionConfig;
 import org.idp.server.core.openid.authentication.config.AuthenticationResponseConfig;
+import org.idp.server.core.openid.authentication.interaction.execution.*;
 import org.idp.server.core.openid.authentication.repository.AuthenticationConfigurationQueryRepository;
 import org.idp.server.platform.datasource.Transaction;
 import org.idp.server.platform.json.JsonNodeWrapper;
